bsp: Retarget Arm Compiler and disable semihosting
Prior to this commit, the BSP retargetted stdio to the USART driver when
building with GNU but used the FVP's semihosting when building with the
Arm Compiler. For consistency, add retarget support for the Arm Compiler
and remove the FVP's run argument for semihosting.

The code is imported from the Open IoT SDK.

Signed-off-by: Lingkai Dong <lingkai.dong@arm.com>