aws-iot-example: Add `appCONFIG_DEVICE_ADVISOR_TEST_ACTIVE`
The config option `appCONFIG_DEVICE_ADVISOR_TEST_ACTIVE` defines
whether device advisor test is active or not. While running device
advisor tests, this config option should must be set to 1. An additional
delay of 8 seconds is added to the backoff delay to ensure that device
advisor moves to the next test before initiate a new connection.

In addition, increase the network and logging buffer size to 20480 to
accommodate the `TLS Receive Maximum Size Fragments` test requirements.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>