credentials-check: Change to run-time check
Modify the AWS client credentials check to be part of the
main application (run-time check) rather than being part of
the compile-time checks.

This change ease the development process because
it is no longer needed to modify the aws_clientcredential.h
file to build the application and then revert it back when
pushing the changes remotely.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>