Use ubuntu 20.04 for formatting so that this repo can use the same formatting as all other FreeRTOS Repos.
Add the formatting bot to this repo.

CMSIS Is not directly included in this repo, removing it from this
repo's manifest.yml file

Fix a broken link in aws-iot-example.md, remove the trailing comma from
a link, and add an exclude for a bash link

Add the link and manifest verification workflow to this repo. Bump
checkout actions to v3 to remove the deprecation warnings.

Signed-off-by: Soren Ptak <ptaksoren@gmail.com>