uncrustify: Use replace option for correcting findings
currently, we are using --check option to check for uncrustify
failures, it is better to use --replace instead of --check as it
corrects the failures beside doing what --check does. Developers can
benefit from this improvement when they run the pre-commit locally. In
addition, use same command line options for uncrustify for both
internal and public CI.

Uncrustify should check for all the FRI project files
except for those that are downloaded with git submodule,
to do so, we have to de-init all the submodules before
Uncrustify tool run and then initialise them back right
after it finish.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>