code-formatting: Remove language override
When performing quality checks with uncrustify,
do not longer override the programming langauge with C.
Instead, let the application detect the language on its own.
This is because C++ files will be added for ML examples
and the C language override for uncrustify introduces
error when it encounters C++ files.

Signed-off-by: Hugues Kamba-Mpiana <hugues.kambampiana@arm.com>