corstone300: Modify linker scripts to run ML applications
This commit is meant to modify Corstone-300 linker files
to be able to run ML applications, the changes are:

- Use both ISRAM0 and ISRAM1 for non-secure data
storage as ML applications would require more RAM space.

- Increase the heap size in the Arm GNU Toolchain linker script
to match Arm Compiler for Embedded's scatter file allocated heap size.
Note that the greater heap size is necessary for Arm Compiler for
Embedded builds of the ML apps.

- Place the neural network models at the DDR space.

- Modify AN552 GNU linker script to match Total Solutions
AN552 GNU linker script as they should be matched to
get ML applications working on Corstone-300

- VSI needs a buffer to transfer audio data from S to NS
space, so in order to provide consistency, VSI
audio buffer would be placed at the DDR space for
Corstone-300 right after where the activation buffer
for SRAM is placed.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>