helpers-sntp: Add helper library for time synchronisation
This helper library shall be used with any FRI application
to add real clock time synchronisation for the target.

This addition is needed to fix AWS TLS Expired Server Cert
test which is part of AWS Device Advisor Validation tests as
this test requires the device under test to have synchronised
real time clock to be able to detect that the server's certificate
is expired and reject it.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>