ml-eval-kit: Make ETHOSU_ARCH configurable in TensorFlow CMake
This patch removes the hardcoded value of u55 in tesorflow.cmake.
It now uses ETHOS_U_NPU_ID as input and convert it to the format
needed by TensorFlow build.
This also means the ETHOS_U_NPU_ID variable preserve workaround is
no longer needed in SetupMlEmbeddedEvaluationKitLibraries.cmake as
it is no longer overwritten in tensorflow.cmake.

Signed-off-by: Gabor Abonyi <gabor.abonyi@arm.com>