ci: Modify test jobs to improve reliability
This commit is meant to do the following:

* Modify test-application-output job to test blinky
application only as it can be used as a smoke test.

* Test ML inference output for both keyword-detection and
speech-recognition applications as part of the OTA test job
which is now renamed to test_applications as it is testing
OTA update and ML inference outputs as well.

Note:
* The mentioned changes would result in a better reliability
and better CI performance.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>