aws-iot-example: Remove example
`aws-iot-example` is deprecated as `keyword-detection`
and `speech-recognition` applications do all the functionalities
that `aws-iot-example` does. Hence, there is no need to maintain
it anymore.

GitHub CI is modified to build `keyword-detection`
and `speech-recognition` applications instead of `aws-iot-example`.

Building the Ml Embedded Evaluation Kit requires Python 3.10
and the `libsndfile1-dev` library.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>