build-run: Use a shell variable to gather CMake arguments
This is so some CMake arguments can be excluded more easily
depending on the build configuration. It will remove the need
to set default values for CMake cache variables from the scripts
as CMake cache variables already have default values.

This change allows a standard overriding of CMake cache variable
in CMake scripts without a check to see if the user input a value
for the cache variable.

Signed-off-by: Gabor Abonyi <gabor.abonyi@arm.com>