bsp: Improve async serial driver
The CMSIS6 has an IRQ driven CMSIS compliant
UART driver.

USART0 is set to non-secure.

TX/RX are now also turned ON at initialization
of the USART.
Because the driver is IRQ driven, the send function
only initiates the transaction. To maintain the CMSIS5
operation model, a busy wait for the completion of the
send operation is required.

Signed-off-by: Dávid Házi <david.hazi@arm.com>