mlek: Increase build speed with reduced use case resource metadata
Patch the library to provide a reduced set of metadata
to download and optimize when building applications.

The only models specified in the use case resources files
(depending on the application) are asr and kws.

Those are the only models required by the FRI applications.

The build time for the ML applications have been
reduced by 40%.

Signed-off-by: Hugues Kamba-Mpiana <hugues.kambampiana@arm.com>