ml-model: Separate model image from NS image
The idea is to move the ML model to the DDR memory region,
generate a separate image for the DDR memory region where
the model can be over the air updated without the need
to update the NS image. To do so the following changes are
needed:

* Modify the linker scripts for all platforms to remove the
RAM load address of the ML model image as it should be
loaded to DDR memory region.

* Some linker sections were renamed to be more descriptive.

* A new CMake function is introduced to extract sections from
input AXF file, this function is used to extract DDR section
which contains the ML model out of the NS image to be loaded
separately.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>