tools: Add short options to build and run scripts
VS Code tasks configuration used for automating building and debugging,
doesn't support optional arguments.

One of the possible solutions is to pass both `command-line option`
(--toolchain) and `value` (GNU) as VS Code tasks `promptString`, but a
space must be added between `command-line option` and `value`. VS Code
encapsulates this `promptString` with a space in double quotes causing
shell not recognising this as a valid command-line option.

Other solution is to introduce a short `command-line option`
(-T, toolchain), so that both `command-line option` and `value` can be
passed in a `promptString` without space. If this is an optional
argument, then an empty string can be passed as `promptString`.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>