tools: Add display command-line option to run script
With the introduction of VS Code build and debug support in FRI using
devcontainers, the development environment becomes headless (no
display). This causes issues with Corstone-315 and object detection
example. Until, we find a solution to X11 forwarding with devcontainers,
a command-line option is provided to disable HDLCD display in
Corstone-315 when running in headless mode.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>