integration-tests: Fix nightly job failures
With https://github.com/FreeRTOS/FreeRTOS-Plus-TCP/commit/4471af64de1d481374881917bbba8736b1217530,
a validation logic has been introduced which checks, if one of the
address in an IP packet is loopback then both source and destination
addresses must be loopback. For network interface tests, we run the
server on the local machine and use loopback interface address in
integration tests to communicate with the server. This causes
validation logic to fail, as the destination address in the IP in the
packet returned by the server is not a loopback address, but instead
address assigned to the device. As a workaround, we use the ethernet
interface address of the host machine instead of loopback to
communicate with the server.

In addition, add a job retry logic to nightly build jobs so that, jobs
gets restarted for known failures.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>