freertos-kernel: Update to `V11.1.0` release
Even though CMake supports circular dependency, but currently it is
supported only for static libraries. If CMake detects a circular
dependency which involves an object library then CMake stops. For more
info, check https://gitlab.kitware.com/cmake/cmake/-/issues/17905.

With FreeRTOS-Kernel 11.1.0 release, FreeRTOS-Kernel produces a single
static library. As part of this effort, `freertos_kernel_port` library
has been changed to an `OBJECT` library. With our current dependency
chain in FRI, this leads to a circular dependency with an object
library. Therefore, remove the dependency of `freertos_config` on
`fri-bsp` to avoid this. Because of this, we've to add the necessary
includes from bsp to `freertos-config`.

The CS315 network interface layer inside the FreeRTOS TCP/IP stack
https://github.com/FreeRTOS/FreeRTOS-Plus-TCP/blob/main/source/portable/NetworkInterface/MPS4_CS315/Device/Include/SSE315.h#L62
is done in such a way that it depends on fri-bsp. Therefore, add
`freertos_plus_tcp` dependency on `fri-bsp`.

In general, a library A  depending on another library B should avoid
linking against granular libraries defined inside the library B. This
can cause build failures in future. Therefore, the dependency of
`isp_platform_driver_system` and `isp_control` on
`freertos_kernel_include` has been changed to `freertos_kernel`.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>