tf-m: Remove change provisioning log message patch
The provisioning message change patch can't be pushed
upstream as it is mainly done to adapt to our CI pass
criteria. The intention is to remove of the out of
tree patches, thus custom pass_output.log files have been
created for each platform as the serial statements on
the terminal are platform specific and should have different
pass_output.log files to define the pass/success criteria for
each platform.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>