mqtt-tests: Refactor mqtt keep alive patch
This patch is modified so it can be pushed upstream
The idea is to check if the `MQTT_KEEP_ALIVE_INTERVAL_SECONDS`
macro is defined by the user. Where if it is not defined,
a default value (5) would be used, this way the user can choose
the default value or different one that is to be provided using
CMake or command line. In our case, target_compile_definitions()
CMake function is used to overwrite the value of the macro for
mqtt integration tests.

This is done to avoid maintaining a patch which is project specific.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>