ml-update-demo: Update from a faulty model to a working one
Add a modified tflite file that produces no inference results
at runtime. This file can be used to build an application that
is fully functional, but the ML inference is unsuccessful in
detecting any keyword. Applying an OTA ML model update in this
state is very visible since the updated model does work as
expected (detects keywords from audio samples).

Signed-off-by: Filip Jagodzinski <filip.jagodzinski@arm.com>
Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>