keyword: Define device and provision keys as volatile
FRI supports PSA crypto implementation either using TF-M on the secure
side or using Mbed TLS on the non-secure side. When using Mbed TLS on
the non-secure side as PSA crypto implementation, the device and
provision keys are defined as volatile since we currently don't have
support for file system.

In addition, when using Mbed TLS on the non-secure side as PSA crypto
implementation, `MBEDTLS_PKCS1_V21` (support for PKCS#1 v2.1 encoding)
is enabled for OTA image validation to work.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>