build: Change default build toolchain to GNU
Currently, the default build toolchain is ARMClang. However, with VS
Code devcontainers, only Arm GNU Toolchain is provided. Therefore, the
default toolchain is changed to GNU. The associated documentation is
also updated. In order to use the Arm Compiler for Embedded (ArmClang),
the toolchain needs to be installed in the devcontainer and an
additional argument `-TARMCLANG` must be passed while building the
example.

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>