tools: Combine createIoTThings.py commands
Introduces a new command to `tools/scripts/createIoTThings.py` that
1. Creates a Thing and Policy.
2. Re-builds the target application (if required).
3. Creates a Bucket, Role, and Update.
This command is documented in `aws_tool.md`.

During application building, the script prints '.' on the same line and
flushes.

The command only takes 2 command line arguments, with the rest
specified in a .json file, an example of which is also included in This
commit as well as the `aws_tool.md` documentation.
Only 3 fields in `aws_tool.md` need to be filled in.

The command also re-uses AWS entities where possible. E.g. if
the Role specified already exists, it checks it can access the Bucket
name provided before re-using the role.
Using Roles not created by your current account is not supported,
as the script may be unable to assume them to check access to the
Bucket.

The motivation for this commit is to make creating OTA updates
easier.

This commit also introduces a new command for cleaning up all
AWS entities specified by the .json config file.
This command is called `cleanup-simplified`.
It will force-delete everything (including bucket contents). This
is not configurable.

Signed-off-by: Reuben Cartwright <Reuben.Cartwright@arm.com>