tools: Fix script overwrite credentials behaviour
This commit makes the script correctly ask before overwriting
credentials.
Previously, the script checked for 'dummy_thingname' and the default
endpoint in 'aws_clientcredential.h' and only asked if these defaults
were not present.
These strings occur in more than one place in the header file so so
this check is not sufficient to say the credentials #define'd have not
been modified by the user.
This commit makes the script check the entirety of the #define line.

Signed-off-by: Reuben Cartwright <Reuben.Cartwright@arm.com>