tools: Extend AWS entity cleanup command
In `createIoTThings.py` there is a command that takes a .json config
and deletes all AWS entities described in there.

This commit extends that command to search the credentials directory,
and identify AWS Thing certificates before deleting these
Things and their possibly related entities (which are generated using
the .json config as well). This behaviour only occurs if the `extended`
flag is set on the cleanup command.

This commit also documents this command in `aws_tool.md`.

Signed-off-by: Reuben Cartwright <Reuben.Cartwright@arm.com>