tools: Support JSON config for all script commands
This commit extends support for the .json config introduced by
`create-update-simplified` and `cleanup-simplified`.
This commit also documents this change in `aws_tool.md` and
updates the example .json config with 4 new fields.
The user-inputted variables are now stored in a dictionary, as this
makes it possible to extract the logic for getting these variables
(either from the JSON or command line) into a single function to avoid
code duplication.
This commit also modifies internal storage of application types to use
'-' rather than '_' between words.

Signed-off-by: Reuben Cartwright <Reuben.Cartwright@arm.com>