docs: Document unit testing with configASSERT
This commit adds an explanation in `unit_testing.md` on how
configASSERT is handled in `freertos_command_pool.c` tests and other
test files.
The intended behaviour is that an assertion failure causes the rest of
the code to stop running, but can be handled by a test and does not
stop the rest of the test's checks running.

Signed-off-by: Reuben Cartwright <Reuben.Cartwright@arm.com>