mps4: Extract the ML model to a separate binary
Eventually, we want the MCUBoot (and the rest of TF-M) to handle the ML
model in the same way as other components. To achieve that, the ML model
component will be kept in flash during boot, for the MCUBoot to validate
the image, and perform the swap operation if needed.

Since the Ethos NPU doesn't have access to flash, the model will be
copied back to DDR at runtime. This is why the model is still kept in
the DDR memory region in the linker script.

Patches for the trusted_firmware-m component:
- Add support for the third image for mps4 platforms.
- Configure the signing layout for the ML model, and reuse the NS key.
- Add a modified flash map for the mps4 platforms, for the MCUBoot to
handle the third image. This is based on the default flash layout
used before.

These changes applies to both:
* Corstone-315 Platform.
* Corstone-320 Platform.

Signed-off-by: Filip Jagodzinski <filip.jagodzinski@arm.com>
Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>