build: TF-M built separately in build.sh
TF-M is built in build.sh and the configuration parameters are passed
for the build in an initial cache script. The TfmInitialCache.cmake
is only added for the blinky application in this commit. After TF-M is
built, the artifacts to be used during the NS build are exported in the
api_ns folder.

Signed-off-by: Gergely Kovacs <Gergely.Kovacs2@arm.com>