fri: Address the issues introduced by Ubuntu update
As a result of updating from Ubuntu v20.04 to v22.04,
the following changes had to be made:

* The default `setuptools_scm` v8.x requires that
`setuptools` package has a version higher than 61.
Hence, updating the minimum required `setuptools`
version.

* The `license_file` is now deprecated and `license_files`
option should be used instead.

* The way CI-env-private installs `uncrustify` has been changed
where it introduced a known issue in the `uncrustify` tool so a
workaround has been applied to overcome the issue.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>