mqtt-agent: Increase MQTT Agent keep alive interval
The `MQTT_AGENT_KEEP_ALIVE_INTERVAL_SECONDS` is increased
to 100 as in case of using SOFTWARE inference with speech
recognition application the connection is dropped before
the ML Inference is done as the software inference is generally
much slower compared to the Ethos NPU. This connection drop
cause the ML inference to have wrong results as the task
is interrupted. Hence, increasing the keep alive interval
to avoid the connection drop.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>