Add OTA Orchestrator fixes (#107)
* ota-orchestrator: Add multiple enhancements and fixes

* Do not suspend OTA agent after successful update.

* vOtaNotActiveHook() should be called whenever there
are no OTA jobs available.

* Use LogDebug and LogError instead of LogInfo to
follow the previous ota-for-embedded-iot-sdk library
implementation (avoiding new issues).

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>
Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>

* mqtt-agent: Increase MQTT Agent keep alive interval

The `MQTT_AGENT_KEEP_ALIVE_INTERVAL_SECONDS` is increased
to 100 as in case of using SOFTWARE inference with speech
recognition application the connection is dropped before
the ML Inference is done as the software inference is generally
much slower compared to the Ethos NPU. This connection drop
cause the ML inference to have wrong results as the task
is interrupted. Hence, increasing the keep alive interval
to avoid the connection drop.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>

* github-upload-artifact: Upgrade the deprecated version

actions/upload-artifact v3 is now deprecated and cannot
be used anymore. Hence, updating to v4.

`strategy.job-index` is post-fixed to the artifact
name to overcome the known issue of upload-artifact v4
not allowing multiple artifacts with the same name on
the workflow run.

Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>

---------

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>
Signed-off-by: Ahmed Ismail <Ahmed.Ismail@arm.com>
Co-authored-by: Devaraj Ranganna <devaraj.ranganna@arm.com>